Выносит createDestinationListTemplate из утилит в темлпейт
