+++ src/view/event-edit-view.js
@@ -1,13 +1,14 @@
 import he from 'he';
 import AbstractStatefulView from '../framework/view/abstract-stateful-view.js';
 import createTypeListTemplate from '../templates/type-list-template.js';
+import createDestinationListTemplate from '../templates/destination-list-template.js';
+
 import {
   humanizePointInputDateTimeType,
   getDestinationName,
   getTypeOffers,
   getDestinationPhotos,
   getDestinationObject,
-  createDestinationListTemplate,
   createDestinationPhotosTemplate,
   createDestinationDescriptionTemplate,
   createOffersSectionTemplateTemplate,
